+++ test/python_raise_error.py
@@ -1,0 +1,2 @@
+def raise_exception():
+  raise Exception('test exception')